requirements.txt
aiohttp==3.7.4.post0
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.5.2
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.9.1
blinker==1.4
certifi==2021.10.8
cffi==1.14.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
comtypes==1.1.11
configparser==5.2.0
cryptography==37.0.2
discord==1.7.3
discord.py==1.7.3
distlib==0.3.4
Django==2.2.12
django-bootstrap4==1.1.1
django-misaka==0.2.1
fastapi==0.78.0
filelock==3.7.0
Flask==2.1.2
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.13.0
h2==4.1.0
houdini.py==0.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.15.0
misaka==2.1.1
multidict==6.0.2
notify2==0.3.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
outcome==1.1.0
packaging==21.3
Pillow==9.1.1
pipenv==2022.5.2
platformdirs==2.5.2
priority==2.0.0
psycopg2==2.9.3
pycparser==2.20
pydantic==1.9.1
Pygments==2.6.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
PyPDF2==1.27.12
pypiwin32==223
PySocks==1.7.1
python-dotenv==0.20.0
python-fsutil==0.6.0
pyttsx3==2.90
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.3.1
toml==0.10.2
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
watchdog==2.1.7
Werkzeug==2.1.2
wsproto==1.1.0
yarl==1.7.2
zipp==3.8.0